[fips] RHEL RPM build
FIPS prohibits `fpm`'s MD5 signature from being accepted so we need to special build the package.
We'll keep it separate for now until we can test SHA256 signatures across enough supported
RPM-based distros to ensure we aren't breaking if we switch to SHA256 for all RPMs.

Also we should update the signature to signal that these are *our* packages and not upstream's.