[common] Increase binary sizes in binarytest
Solo filters and changes increase the size of some of our binaries. We should increase the max to account for this, but retain the test to ensure we don't end up smaller (thus something is missing).

At some point we should look at these values to find more reasonable maxes to look for. If we care?

Signed-off-by: Daniel Hawton <daniel.hawton@solo.io>